sort all the materials into different catagories
