--- src/main/java/com/nomiceu/nomilabs/item/unification/LabsElements.java
@@ -1,11 +1,0 @@
-package com.nomiceu.nomilabs.item.unification;
-
-import gregtech.api.unification.Element;
-import gregtech.api.unification.Elements;
-
-public class LabsElements {
-    public static final Element Nm = Elements.add(130, 234, -1, null,"Omnium", "Nm", false);
-    public static final Element Dc = Elements.add(149, 264, -1, null, "Draconium", "Dc", false);
-    public static final Element ADc = Elements.add(149, 264, -1, null, "AwakenedDraconium", "Dc*", false);
-    public static final Element Tn = Elements.add(43, 55, -1, null, "Taranium", "Tn", false);
-}
